+++ dart_sample/lib/sample_routes.dart
@@ -92,6 +92,7 @@ class SampleRoutes {
   static const String overflowIndicator = '/overflow-indicator';
   static const String offstage = '/offstage';
   static const String baseline = '/baseline';
+  static const String layoutBuilder = '/layout-builder';
 }
 
 class SampleRouteDefinition {